+++ include/appinfo.h
@@ -8,13 +8,16 @@
 #define APPINFO_H
 
 /***
- * @brief Compiler version
+ * @brief Compiler current version
  * @warning Dont forget to update CHANGELOG.md, debian/changelog and
  * rpmbuild/SPECS/msxbas2rom.spec files
- * @note Run the code below to create a new release on github:
+ * @remark Run the code below to create a new release on github:
  * git tag -a v0.0.0.0 -m "Release 0.0.0.0" && git push msxbas2rom --tags
+ * git checkout -b release/v0.0.0.0
+ * git push msxbas2rom release/v0.0.0.0
+ * git switch master
  */
-#define app_version "0.3.3.1"
+#define app_version "0.3.3.2"
 
 //! @brief Header information
 const char *info_header = R"(MSXBAS2ROM - MSX BASIC TO ROM COMPILER
@@ -75,6 +78,8 @@ https://github.com/amaurycarvalho/msxbas2rom/wiki/Usage
 const char *info_history = R"(
 Version history
 
+0.3.3.2 – (2025/##/##) Symbols/resources refactoring and NoICE 
+                       OpenMSX Debugger support;
 0.3.3.1 – (2025/09/03) Symbols/resources refactoring, OMDS support,
                        CLEAR statement and minor bug fixes;
 0.3.3.0 – (2025/08/18) New -i and -o parameters;