+++ src/compiler.cpp
@@ -80,7 +80,7 @@ bool Compiler::build(Parser* parser) {
   codeItem->length = code_pointer - codeItem->start;
   codeItem->is_code = true;
   codeItem->debug = true;
-  resourceManager.codeList.push_back(codeItem);
+  symbolManager.codeList.push_back(codeItem);
   if (opts->debug) printf(" %i byte(s)\n", codeItem->length);
 
   if (opts->debug) printf("Registering start of program...");
@@ -92,7 +92,7 @@ bool Compiler::build(Parser* parser) {
   codeItem->length = code_pointer - codeItem->start;
   codeItem->is_code = true;
   codeItem->debug = true;
-  resourceManager.codeList.push_back(codeItem);
+  symbolManager.codeList.push_back(codeItem);
   if (opts->debug) printf(" %i byte(s)\n", codeItem->length);
   if (codeItem->length >= 0x4000) {
     syntax_error("Maximum of start of program code per ROM reached (16k)");
@@ -144,7 +144,7 @@ bool Compiler::build(Parser* parser) {
       codeItem->length = code_pointer - codeItem->start;
       codeItem->is_code = true;
       codeItem->debug = true;
-      resourceManager.codeList.push_back(codeItem);
+      symbolManager.codeList.push_back(codeItem);
 
       if (opts->debug) printf("/%i", codeItem->length);
 
@@ -178,7 +178,7 @@ bool Compiler::build(Parser* parser) {
     codeItem->length = code_pointer - codeItem->start;
     codeItem->is_code = true;
     codeItem->debug = true;
-    resourceManager.codeList.push_back(codeItem);
+    symbolManager.codeList.push_back(codeItem);
     if (opts->debug) printf(" %i byte(s)\n", codeItem->length);
     if (codeItem->length >= 0x4000) {
       syntax_error("Maximum of end of program code per ROM reached (16k)");
@@ -193,7 +193,7 @@ bool Compiler::build(Parser* parser) {
     codeItem->length = code_pointer - codeItem->start;
     codeItem->is_code = true;
     codeItem->debug = false;
-    resourceManager.codeList.push_back(codeItem);
+    symbolManager.codeList.push_back(codeItem);
     if (opts->debug) printf(" %i byte(s)\n", codeItem->length);
     if (codeItem->length >= 0x4000) {
       syntax_error("Maximum of support code per ROM reached (16k)");
@@ -258,6 +258,7 @@ void Compiler::clear_symbols() {
   symbols.clear();
   fixes.clear();
 
+  symbolManager.clear();
   resourceManager.clear();
 
   while (!forNextStack.empty()) forNextStack.pop();
@@ -331,7 +332,7 @@ int Compiler::save_symbols() {
             codeItem->length = code_pointer - codeItem->start;
             codeItem->is_code = false;
             codeItem->debug = true;
-            resourceManager.codeList.push_back(codeItem);
+            symbolManager.codeList.push_back(codeItem);
 
             length += codeItem->length;
           }
@@ -343,7 +344,7 @@ int Compiler::save_symbols() {
           codeItem->addr_within_segm = ram_page + ram_pointer;
           codeItem->is_code = false;
           codeItem->debug = true;
-          resourceManager.dataList.push_back(codeItem);
+          symbolManager.dataList.push_back(codeItem);
 
           var_size = 0;
 
@@ -813,7 +814,7 @@ int Compiler::write(unsigned char* dest, int start_address) {
 
   // copy compiled code to final destination
 
-  t = resourceManager.codeList.size();
+  t = symbolManager.codeList.size();
   addr_within_segm = start_address;
 
   if (opts->megaROM) {
@@ -826,7 +827,7 @@ int Compiler::write(unsigned char* dest, int start_address) {
     d = dest;
 
     for (i = 0; i < t; i++) {
-      codeItem = resourceManager.codeList[i];
+      codeItem = symbolManager.codeList[i];
 
       // printf("%i address %i size %i\n", i, codeItem->start,
       // codeItem->length);
@@ -911,7 +912,7 @@ int Compiler::write(unsigned char* dest, int start_address) {
 
   } else {
     for (i = 0; i < t; i++) {
-      codeItem = resourceManager.codeList[i];
+      codeItem = symbolManager.codeList[i];
       codeItem->segm = 0;
       codeItem->addr_within_segm = addr_within_segm;
       addr_within_segm += codeItem->length;